+++ src/main/java/kr/ac/jbnu/se/tetris/controllers/TetrisFrameController.java
@@ -1,14 +1,15 @@
-package kr.ac.jbnu.se.tetris.views;
+package kr.ac.jbnu.se.tetris.controllers;
 import kr.ac.jbnu.se.tetris.models.KeyInput;
 import kr.ac.jbnu.se.tetris.models.Member;
+import kr.ac.jbnu.se.tetris.controllers.PlayerPageController;
 
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.KeyEvent;
 
 
-public class TetrisFrame extends JFrame {
-    public TetrisFrame() {
+public class TetrisFrameController extends JFrame {
+    public TetrisFrameController() {
     }
 
     public void init() {
@@ -16,11 +17,9 @@ public class TetrisFrame extends JFrame {
         setLayout(new BorderLayout());
         setPreferredSize(new Dimension(500, 800));
 
-        KeyInput p1Key = new KeyInput("data/player1key.json");
-        PlayerPage playerPage = new PlayerPage(new Member(), p1Key);
+        KeyInput p1Key = new KeyInput("src/main/java/kr/ac/jbnu/se/tetris/data/player1key.json");
+        PlayerPageController playerPage = new PlayerPageController(new Member(), p1Key);
         add(playerPage,BorderLayout.CENTER);
-
-        playerPage.init();
 
         setTitle("Tetris");
         setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
